refactor(tests): extract inline tests and fix 6 bugs
Bug fixes:
- zygote-rs: exit_status key collision (HashMap->HashSet)
- image-puller: RefCell->Mutex for async Send
- network-rs: nested tokio runtime panic fix
- cgroups-rs: CgroupChoice enum test compile fix
- integration-tests: AdaptiveEngine::new() Result API fix
- phantom-cli: run.rs pool.execute() return type fix

Test extraction:
- Extract 98 inline test functions from 27 src files to separate tests/ dirs
- 23 crates now have zero inline #[cfg(test)] blocks
- 26 new test files created across core, storage, security, monitoring, tools
- 46+ tests pass, 6 ignored (privileged/SIGSEGV/known issues)

Security:
- Revert pub(crate) exposure in mcp crate (ALLOWED_COMMANDS, sanitize_command)
- Restore inline tests for pub(crate) security internals (only legitimate exception)

Docs:
- Add 3 mandatory rules to AGENTS.md (sub-agent only, separate test files, precision)
