+++ src/cli/phantom-cli/src/fragment_registry.rs
@@ -285,118 +285,3 @@ impl FragmentRegistry {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use tempfile::TempDir;
-
-    #[test]
-    fn test_registry_create_and_list() {
-        let temp_dir = TempDir::new().unwrap();
-        let registry_path = temp_dir.path().join("registry.json");
-        let mut registry = FragmentRegistry::with_path(&registry_path).unwrap();
-
-        let info = FragmentInfo {
-            name: "test-fragment".to_string(),
-            profile: "sandbox".to_string(),
-            pid: Some(12345),
-            created_at: 1234567890,
-            status: FragmentStatus::Running,
-            components: vec!["tcp".to_string(), "dns".to_string()],
-            memory_kb: 8192,
-            cpu_count: None,
-            mode: "sandbox".to_string(),
-            last_validated: None,
-            image: None,
-        };
-
-        registry.register(info.clone()).unwrap();
-
-        let listed = registry.list();
-        assert_eq!(listed.len(), 1);
-        assert_eq!(listed[0].name, "test-fragment");
-    }
-
-    #[test]
-    fn test_registry_persistence() {
-        let temp_dir = TempDir::new().unwrap();
-        let registry_path = temp_dir.path().join("registry.json");
-
-        {
-            let mut registry = FragmentRegistry::with_path(&registry_path).unwrap();
-            let info = FragmentInfo {
-                name: "persistent-fragment".to_string(),
-                profile: "direct".to_string(),
-                pid: Some(99999),
-                created_at: 1234567890,
-                status: FragmentStatus::Running,
-                components: vec![],
-                memory_kb: 4096,
-                cpu_count: None,
-                mode: "direct".to_string(),
-                last_validated: None,
-                image: None,
-            };
-            registry.register(info).unwrap();
-        }
-
-        // Reload and verify
-        let registry = FragmentRegistry::with_path(&registry_path).unwrap();
-        assert_eq!(registry.list().len(), 1);
-        assert_eq!(
-            registry.get("persistent-fragment").unwrap().pid,
-            Some(99999)
-        );
-    }
-
-    #[test]
-    fn test_cleanup_old() {
-        let temp_dir = TempDir::new().unwrap();
-        let registry_path = temp_dir.path().join("registry.json");
-        let mut registry = FragmentRegistry::with_path(&registry_path).unwrap();
-
-        let now = SystemTime::now()
-            .duration_since(SystemTime::UNIX_EPOCH)
-            .unwrap()
-            .as_secs();
-
-        // Add old stopped fragment
-        let old_info = FragmentInfo {
-            name: "old-fragment".to_string(),
-            profile: "sandbox".to_string(),
-            pid: None,
-            created_at: now - 7200, // 2 hours ago
-            status: FragmentStatus::Stopped,
-            components: vec![],
-            memory_kb: 4096,
-            cpu_count: None,
-            mode: "sandbox".to_string(),
-            last_validated: None,
-            image: None,
-        };
-        registry.register(old_info).unwrap();
-
-        // Add recent running fragment
-        let recent_info = FragmentInfo {
-            name: "recent-fragment".to_string(),
-            profile: "sandbox".to_string(),
-            pid: Some(12345),
-            created_at: now - 60, // 1 minute ago
-            status: FragmentStatus::Running,
-            components: vec![],
-            memory_kb: 4096,
-            cpu_count: None,
-            mode: "sandbox".to_string(),
-            last_validated: None,
-            image: None,
-        };
-        registry.register(recent_info).unwrap();
-
-        // Clean up fragments older than 1 hour
-        let removed = registry.cleanup_old(3600).unwrap();
-        assert_eq!(removed, 1);
-        assert_eq!(registry.list().len(), 1);
-        assert!(registry.exists("recent-fragment"));
-        assert!(!registry.exists("old-fragment"));
-    }
-}
