+++ src/core/network-rs/Cargo.toml
@@ -14,3 +14,6 @@ tokio = { version = "1", features = ["full"] }
 serde = { workspace = true }
 ipnetwork = "0.20"
 thiserror = { workspace = true }
+
+[dev-dependencies]
+tokio = { version = "1", features = ["full", "test-util", "macros", "rt"] }
