+++ src/core/types/types-rs/Cargo.toml
@@ -12,5 +12,8 @@ serde = { workspace = true }
 thiserror = { workspace = true }
 libc = { workspace = true }
 
+[dev-dependencies]
+serde_json = "1.0"
+
 [build-dependencies]
 cbindgen = { workspace = true }
