+++ src/core/types/types-rs/src/lib.rs
@@ -452,3 +452,4 @@ pub struct ExecutionResult {
     pub output: String,
     pub exit_code: i32,
 }
+
