+++ src/core/types/types-rs/tests/mod.rs
@@ -1,0 +1,3 @@
+pub mod errors;
+pub mod ffi;
+pub mod metadata;
