+++ src/security/cgroups-rs/src/lib.rs
@@ -541,48 +541,3 @@ pub fn is_in_user_slice() -> bool {
     false
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_cgroup_manager_paths() {
-        let manager = CgroupManager::new("test_container");
-        assert_eq!(
-            manager.full_path,
-            PathBuf::from("/sys/fs/cgroup/test_container")
-        );
-
-        let manager_abs = CgroupManager::new("/tmp/cgroup/test");
-        assert_eq!(manager_abs.full_path, PathBuf::from("/tmp/cgroup/test"));
-    }
-
-    #[test]
-    fn test_user_slice_status_enum() {
-        // Test that all variants exist and are distinct
-        assert_ne!(UserSliceStatus::Available, UserSliceStatus::Unavailable);
-        assert_ne!(UserSliceStatus::CanCreate, UserSliceStatus::NotWritable);
-    }
-
-    #[test]
-    fn test_get_user_slice_path() {
-        let path = get_user_slice_path();
-        let uid = unsafe { libc::getuid() };
-        let expected = format!("/sys/fs/cgroup/user.slice/user-{}.slice", uid);
-        assert_eq!(path, expected);
-    }
-
-    #[test]
-    fn test_detect_cgroup_version() {
-        let version = detect_cgroup_version();
-        // Should return 0, 1, or 2
-        assert!(version <= 2);
-    }
-
-    #[test]
-    fn test_cgroup_choice_default() {
-        // Test that CgroupChoice variants are distinct
-        assert_ne!(CgroupChoice::UseSudo, CgroupChoice::ContinueWithoutCgroups);
-        assert_ne!(CgroupChoice::ConfigureDelegation, CgroupChoice::Abort);
-    }
-}
