+++ src/storage/oci-registry-rs/src/pusher.rs
@@ -506,15 +506,3 @@ impl RegistryPusher {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_push_config_default() {
-        let config = PushConfig::default();
-        assert_eq!(config.max_concurrent, 3);
-        assert!(config.compress);
-    }
-}
