+++ src/tools/integration-tests/src/lib.rs
@@ -36,7 +36,7 @@ mod tests {
         assert!(required.components.contains(&Component::TcpStack));
 
         // 2. Determine Execution Mode
-        let engine = AdaptiveEngine::new();
+        let engine = AdaptiveEngine::new().expect("Failed to create AdaptiveEngine");
         let risk = RiskProfile {
             network_access: true,
             ..Default::default()
