+++ src/tools/phantom-build/src/parser.rs
@@ -347,40 +347,3 @@ pub struct BuildStage {
     pub instructions: Vec<Instruction>,
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_parse_simple_fragmentfile() {
-        let content = r#"
-FROM rust:1.75-alpine
-WORKDIR /app
-COPY . .
-RUN cargo build --release
-CMD ["./target/release/app"]
-"#;
-
-        let fragmentfile = Fragmentfile::from_string(content).unwrap();
-        assert_eq!(fragmentfile.instructions.len(), 5);
-    }
-
-    #[test]
-    fn test_parse_multistage() {
-        let content = r#"
-FROM rust:1.75 AS builder
-WORKDIR /build
-COPY . .
-RUN cargo build --release
-
-FROM alpine:3.18
-COPY --from=builder /build/target/release/app /usr/local/bin/
-CMD ["app"]
-"#;
-
-        let fragmentfile = Fragmentfile::from_string(content).unwrap();
-        let stages = fragmentfile.stages();
-        assert_eq!(stages.len(), 2);
-        assert_eq!(stages[0].name, Some("builder".to_string()));
-    }
-}
